requirements.txt
Flask==2.1.0
requests==2.27.1
opennsfw2==0.10.2
Pillow==9.0.0
Flask-Limiter==3.3.0